Cleaned up code, got a blockade for the player to get up to, next to add the magical plant that they need to climb! :D WOO lets do this (**) also cleaned up code a bit, and removed a few comments and gave a general heading in some parts.

+++ PGAG_ASS/PGAG_ASS/Enemy.h
@@ -9,33 +9,33 @@ public:
 		unsigned int current = SDL_GetTicks();
 		SDL_Rect RenderSize;
 		SDL_Rect TextureSize;
-		float temp = (current / 100) % 6;
-		float temp2 = (current / 100) % 4;
-		float temp3 = (current / 100) % 3;
+		float temp = float((current / 100) % 6);
+		float temp2 = float((current / 100) % 4);
+		float temp3 = float((current / 100) % 3);
 		if (movingL){
-			RenderSize.x = temp *100;
+			RenderSize.x = int(temp * 100);
 			RenderSize.y = 440;
 		}
 		if (movingR){
-			RenderSize.x = temp * 100;
+			RenderSize.x = int(temp * 100);
 			RenderSize.y = 70;
 		}
 		if (attackL){
-			RenderSize.x = temp2 * 100;
+			RenderSize.x = int(temp2 * 100);
 			RenderSize.y =135;
 		}
 		if (attackL){
-			RenderSize.x = temp2 * 100;
+			RenderSize.x = int(temp2 * 100);
 			RenderSize.y = 500;
 		}
 		if (idle){
-			RenderSize.x = temp3 * 100;
+			RenderSize.x = int(temp3 * 100);
 			RenderSize.y = 0;
 		}
 		RenderSize.h = height;
 		RenderSize.w = width;
-		TextureSize.x = Texturesize.x;
-		TextureSize.y = Texturesize.y;
+		TextureSize.x = int(Texturesize.x);
+		TextureSize.y = int(Texturesize.y);
 		TextureSize.h = height;
 		TextureSize.w = width;
 		SDL_RenderCopy(r, Texture::getText(), &RenderSize, &TextureSize);
